i18n. Add Limits functionality. Changes to INSTALL and HACKING documents.
(cherry picked from commit 0a0af1e9cded313ac4d5293d16cf834346287ccb)

+++ libreplan-business/src/main/java/org/libreplan/business/common/entities/Limits.java
@@ -34,7 +34,7 @@ public class Limits extends BaseEntity{
 
     private String type;
 
-    private Long value;
+    private Integer value;
 
 
     public String getType() {
@@ -44,10 +44,10 @@ public class Limits extends BaseEntity{
         this.type = type;
     }
 
-    public Long getValue() {
+    public Integer getValue() {
         return value;
     }
-    public void setValue(Long value) {
+    public void setValue(Integer value) {
         this.value = value;
     }
 }
